requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sentence-transformers==2.2.0
faiss-cpu==1.7.4
anthropic==0.7.0
python-dotenv==1.0.0
pydantic==2.4.0
pytest==7.4.0
python-multipart==0.0.7
pydantic-settings==2.1.0
pypdf>=4.0.0
tqdm>=4.66.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
